requirements.txt
# YOLO (Ultralytics) - detección de vehículos y de matrículas
ultralytics>=8.0.0
# Descarga del modelo de matrículas desde Hugging Face (opcional si ya tienes el .pt)
huggingface_hub>=0.20.0

# PaddleOCR - lectura de matrículas
paddlepaddle>=2.5.0
paddleocr>=2.7.0
# EasyOCR - lectura de matrículas (sin cuDNN)
easyocr>=1.7.0
# Tesseract (test_ocr.py) - requiere también: apt install tesseract-ocr
pytesseract>=0.3.10
# platerec (opcional): pip install "platerec[cpu]" — backend ONNX para matrículas
# fast-alpr (opcional): pip install "fast-alpr[onnx]" — detección + OCR en uno (ONNX)
fast-alpr[onnx]>=0.3.0

# Clasificador marca/modelo (opcional): pip install torch timm — Jordo23/vehicle-classifier
# torch>=2.0.0
# timm>=0.9.0

# Google Sheets - registro de matrículas detectadas (opcional)
gspread>=5.12.0
google-auth>=2.23.0

# Cloudinary - almacenamiento de imágenes (opcional)
cloudinary>=1.41.0

# Imagen y utilidades
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0